+++ DietLog-MVVM/View/SignIn/SignInViewController.swift
@@ -11,16 +11,8 @@ import RxSwift
 
 class SignInViewController: BaseViewController {
     
-    // MARK: - UI Componet
-    private lazy var backgroundView = UIView()
     private lazy var dietLogLabel = UILabel()
-    private lazy var titleLabel = UILabel()
-    private lazy var subTitleLabel = UILabel()
-    private lazy var nicknameTextField = UITextField()
-    private lazy var emailTextField = UITextField()
-    private lazy var passwordTextField = UITextField()
-    private lazy var doneButton = UIButton()
-    private lazy var visibilityToggleButton = UIButton(type: .custom)
+    private lazy var signInView = SignInView()
     private lazy var stackVew = UIStackView()
     private lazy var snsTitleLabel = UILabel()
 
@@ -42,26 +34,10 @@ class SignInViewController: BaseViewController {
     
     // MARK: - Setup UI
     override func setupUI() {
-        view.addSubviews([dietLogLabel, backgroundView, snsTitleLabel, stackVew])
-        
-        backgroundView.addSubviews([titleLabel,
-                                    subTitleLabel,
-                                    nicknameTextField,
-                                    emailTextField,
-                                    passwordTextField,
-                                    doneButton])
-        
-        setupBackgroundViewUI()
+        view.addSubviews([dietLogLabel, signInView, snsTitleLabel, stackVew])
+        signInView.configure()
         setupLabelslUI()
-        setupTextFieldsUI()
-        setupDoneButtonUI()
         setupStackView()
-    }
-    
-    private func setupBackgroundViewUI() {
-        backgroundView.backgroundColor = .white
-        backgroundView.applyShadow()
-        backgroundView.applyRadius()
     }
     
     private func setupLabelslUI() {
@@ -69,50 +45,10 @@ class SignInViewController: BaseViewController {
         dietLogLabel.textAlignment = .center
         dietLogLabel.textColor = .customGreen
         
-        titleLabel.configure(text: SignInText.title, font: .title)
-        titleLabel.textAlignment = .left
-        
-        subTitleLabel.configure(text: SignInText.subTitle, font: .body)
-        subTitleLabel.textAlignment = .left
-        
         snsTitleLabel.configure(text: SignInText.sns, font: .boldBody)
         snsTitleLabel.textAlignment = .center
     }
-    
-    private func setupTextFieldsUI() {
-        nicknameTextField.configure()
-        nicknameTextField.placeholder = SignInText.nickname
-        
-        emailTextField.configure()
-        emailTextField.placeholder = SignInText.email
-        
-        passwordTextField.configure()
-        passwordTextField.placeholder = SignInText.password
-        passwordTextField.isSecureTextEntry = true
-        
-        var buttonConfiguration = UIButton.Configuration.plain()
-        buttonConfiguration.imagePadding = 4
-        buttonConfiguration.baseBackgroundColor = .clear
-        buttonConfiguration.baseForegroundColor = .systemGray4
-        
-        visibilityToggleButton.setImage(.eyeClosed, for: .normal)
-        visibilityToggleButton.setImage(.eyeOpen, for: .selected)
-        visibilityToggleButton.configuration = buttonConfiguration
 
-        passwordTextField.rightView = visibilityToggleButton
-        passwordTextField.rightViewMode = .always
-    }
-    
-    private func setupDoneButtonUI() {
-        let width: CGFloat = CGFloat(ComponentSize.textFieldHeight.rawValue)
-        doneButton.configureCircleShape(width: width)
-        
-        let buttonImgae = UIImage(systemName: ImageName.rightArrow,
-                                  withConfiguration: UIImage.SymbolConfiguration(pointSize: width / 2, weight: .medium))
-        doneButton.setImage(buttonImgae, for: .normal)
-        doneButton.tintColor = .white
-    }
-    
     private func setupStackView() {
         stackVew.axis = .vertical
         stackVew.spacing = 0
@@ -148,7 +84,7 @@ class SignInViewController: BaseViewController {
             make.centerY.equalTo(topView)
         }
         
-        backgroundView.snp.makeConstraints { make in
+        signInView.snp.makeConstraints { make in
             make.centerY.equalToSuperview()
             make.leading.trailing.equalToSuperview().inset(Padding.leftRightSpacing.rawValue)
             make.height.equalTo(view.snp.height).dividedBy(2)
@@ -156,46 +92,13 @@ class SignInViewController: BaseViewController {
         
         dietLogLabel.snp.makeConstraints { make in
             make.centerX.equalToSuperview()
-            make.bottom.equalTo(backgroundView.snp.top).offset(-48)
+            make.bottom.equalTo(signInView.snp.top).offset(-48)
         }
-        
-        titleLabel.snp.makeConstraints { make in
-            make.top.leading.trailing.equalToSuperview().inset(Padding.leftRightSpacing.rawValue)
-        }
-        
-        subTitleLabel.snp.makeConstraints { make in
-            make.top.equalTo(titleLabel.snp.bottom).offset(4)
-            make.leading.trailing.equalTo(titleLabel)
-        }
-        
-        nicknameTextField.snp.makeConstraints { make in
-            make.bottom.equalTo(emailTextField.snp.top).offset(-20)
-            make.leading.trailing.equalTo(titleLabel)
-            make.height.equalTo(ComponentSize.textFieldHeight.rawValue)
-        }
-        
-        emailTextField.snp.makeConstraints { make in
-            make.centerY.equalToSuperview()
-            make.leading.trailing.equalTo(titleLabel)
-            make.height.equalTo(ComponentSize.textFieldHeight.rawValue)
-        }
-        
-        passwordTextField.snp.makeConstraints { make in
-            make.top.equalTo(emailTextField.snp.bottom).offset(20)
-            make.leading.trailing.equalTo(titleLabel)
-            make.height.equalTo(ComponentSize.textFieldHeight.rawValue)
-        }
-        
-        doneButton.snp.makeConstraints { make in
-            make.top.equalTo(passwordTextField.snp.bottom).offset(24)
-            make.centerX.equalToSuperview()
-            make.width.height.equalTo(ComponentSize.textFieldHeight.rawValue)
-        }
-        
+
         snsTitleLabel.snp.makeConstraints { make in
-            make.top.equalTo(backgroundView.snp.bottom).offset(24)
+            make.top.equalTo(signInView.snp.bottom).offset(24)
             make.bottom.equalTo(stackVew.snp.top).offset(-8)
-            make.leading.trailing.equalTo(titleLabel)
+            make.leading.trailing.equalTo(24)
         }
         
         stackVew.snp.makeConstraints { make in
@@ -205,19 +108,19 @@ class SignInViewController: BaseViewController {
     
     // MARK: - Setup Event
     override func setupEvent() {
-        visibilityToggleButton.addTarget(self, action: #selector(toggleVisibilityButton), for: .touchUpInside)
+//        visibilityToggleButton.addTarget(self, action: #selector(toggleVisibilityButton), for: .touchUpInside)
     }
     
     // MARK: - Setup Bind
     override func setupBinding() {
-        doneButton.rx.tap
-            .withLatestFrom(nicknameTextField.rx.text)
-            .map {
-                $0?.trimmingCharacters(in: .whitespacesAndNewlines)
-            }
-            .subscribe() { [weak self] nickname in
-                self?.viewModel.nickname.onNext(nickname)
-            }.disposed(by: disposeBag)
+//        doneButton.rx.tap
+//            .withLatestFrom(nicknameTextField.rx.text)
+//            .map {
+//                $0?.trimmingCharacters(in: .whitespacesAndNewlines)
+//            }
+//            .subscribe() { [weak self] nickname in
+//                self?.viewModel.nickname.onNext(nickname)
+//            }.disposed(by: disposeBag)
         
         
         viewModel.signInResult
@@ -246,8 +149,8 @@ extension SignInViewController {
     }
     
     @objc func toggleVisibilityButton() {
-        passwordTextField.isSecureTextEntry.toggle()
-        visibilityToggleButton.isSelected.toggle()
+//        passwordTextField.isSecureTextEntry.toggle()
+//        visibilityToggleButton.isSelected.toggle()
     }
     
     @objc func loginWithKakao() {
